# backend/docs/swagger.yaml
definitions:
  main.Card:
    properties:
      color:
        type: string
      id:
        type: integer
      text:
        type: string
    type: object
info:
  contact: {}
paths:
  /blackcard:
    get:
      description: Retrieve the current black card
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Card'
      summary: Get current black card
      tags:
      - game
  /card/play:
    post:
      description: Play a card by providing the card index in the header
      parameters:
      - description: Index of the card to play
        in: header
        name: CardIndex
        required: true
        type: integer
      responses:
        "202":
          description: Card accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Player plays a card
      tags:
      - game
  /cards/played:
    get:
      description: Retrieve all the cards that have been played
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Card'
            type: array
      summary: Get all played cards
      tags:
      - game
  /hand:
    get:
      description: Get the hand of the player
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Card'
            type: array
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Get Hand of the player
      tags:
      - player
  /start:
    post:
      description: Start the game
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Start the game
      tags:
      - game
  /ws:
    get:
      description: Establish a WebSocket connection for a player.
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Connect a player via WebSocket
      tags:
      - websocket
swagger: "2.0"